:tractor: Move loader into its own file
