+++ pkg/cmds/sql.go
@@ -5,10 +5,8 @@ import (
 	"fmt"
 	clay_sql "github.com/go-go-golems/clay/pkg/sql"
 	"github.com/go-go-golems/glazed/pkg/cmds"
-	"github.com/go-go-golems/glazed/pkg/cmds/alias"
 	"github.com/go-go-golems/glazed/pkg/cmds/layers"
 	"github.com/go-go-golems/glazed/pkg/cmds/layout"
-	"github.com/go-go-golems/glazed/pkg/cmds/loaders"
 	"github.com/go-go-golems/glazed/pkg/cmds/parameters"
 	"github.com/go-go-golems/glazed/pkg/middlewares"
 	"github.com/go-go-golems/glazed/pkg/settings"
@@ -17,7 +15,6 @@ import (
 	"github.com/pkg/errors"
 	"gopkg.in/yaml.v3"
 	"io"
-	"io/fs"
 	"strings"
 )
 
@@ -255,72 +252,4 @@ func (s *SqlCommand) RunQueryIntoGlaze(
 	gp middlewares.Processor) error {
 
 	return clay_sql.RunQueryIntoGlaze(ctx, db, s.renderedQuery, []interface{}{}, gp)
-}
-
-type SqlCommandLoader struct {
-	DBConnectionFactory DBConnectionFactory
-}
-
-var _ loaders.FileCommandLoader = (*SqlCommandLoader)(nil)
-
-func (scl *SqlCommandLoader) LoadCommandsFromReader(
-	r io.Reader,
-	options []cmds.CommandDescriptionOption,
-	aliasOptions []alias.Option,
-) ([]cmds.Command, error) {
-	return loaders.LoadCommandOrAliasFromReader(
-		r,
-		scl.loadSqlCommandFromReader,
-		options,
-		aliasOptions)
-}
-
-func (scl *SqlCommandLoader) IsFileSupported(f fs.FS, fileName string) bool {
-	return strings.HasSuffix(fileName, ".yaml") || strings.HasSuffix(fileName, ".yml")
-}
-
-func (scl *SqlCommandLoader) loadSqlCommandFromReader(
-	s io.Reader,
-	options []cmds.CommandDescriptionOption,
-	_ []alias.Option,
-) ([]cmds.Command, error) {
-	scd := &SqlCommandDescription{}
-	err := yaml.NewDecoder(s).Decode(scd)
-	if err != nil {
-		return nil, err
-	}
-
-	options_ := []cmds.CommandDescriptionOption{
-		cmds.WithShort(scd.Short),
-		cmds.WithLong(scd.Long),
-		cmds.WithFlags(scd.Flags...),
-		cmds.WithArguments(scd.Arguments...),
-		cmds.WithLayers(scd.Layers...),
-		cmds.WithLayout(&layout.Layout{
-			Sections: scd.Layout,
-		}),
-	}
-	options_ = append(options_, options...)
-
-	sq, err := NewSqlCommand(
-		cmds.NewCommandDescription(
-			scd.Name,
-		),
-		WithDbConnectionFactory(scl.DBConnectionFactory),
-		WithQuery(scd.Query),
-		WithSubQueries(scd.SubQueries),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, option := range options_ {
-		option(sq.Description())
-	}
-
-	if !sq.IsValid() {
-		return nil, errors.New("Invalid command")
-	}
-
-	return []cmds.Command{sq}, nil
 }
